imap: move over STATUS types from imapclient
